Renamed test to match their work. Working on Facade tests
